server/requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.7
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
debugpy==1.6.2
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-pure-pagination==0.3.0
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-jwt==1.19.2
drf-nested-routers==0.93.4
drf-yasg==1.21.3
entrypoints==0.4
environ==1.0
executing==0.9.1
factory-boy==3.2.1
Faker==14.0.0
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
inflection==0.5.1
ipykernel==6.15.1
ipython==8.4.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
kombu==5.2.4
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
msgpack==1.0.4
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.18.3
uvloop==0.16.0
vine==5.0.0
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
